+++ app/src/main/java/com/hongstudio/flabrecyclerviewassignment/view/NormalItemViewHolder.kt
@@ -1,7 +1,8 @@
-package com.hongstudio.flabrecyclerviewassignment
+package com.hongstudio.flabrecyclerviewassignment.view
 
 import androidx.recyclerview.widget.RecyclerView.ViewHolder
 import com.hongstudio.flabrecyclerviewassignment.databinding.ItemNormalBinding
+import com.hongstudio.flabrecyclerviewassignment.model.Item
 
 class NormalItemViewHolder(
     private val binding: ItemNormalBinding,
